@@ -9,17 +9,16 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "@vue/compat": "^3.1.0",
     "axios": "^0.27.2",
     "bootstrap": "^5.1.3",
     "core-js": "^3.22.7",
     "downloadjs": "^1.4.7",
     "vue": "^3.1.0",
-    "vue-router": "^3.5.4",
-    "vuex": "^3.6.2"
+    "vue-router": "^4.0.15",
+    "vuex": "^4.0.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.16",
+    "@babel/core": "^7.18.2",
     "@babel/eslint-parser": "^7.12.16",
     "@vue/cli-plugin-babel": "^5.0.4",
     "@vue/cli-plugin-eslint": "^5.0.4",
@@ -27,16 +26,15 @@
     "@vue/cli-plugin-unit-jest": "^5.0.4",
     "@vue/cli-service": "^5.0.4",
     "@vue/compiler-sfc": "^3.1.0",
-    "@vue/eslint-config-prettier": "^6.0.0",
-    "@vue/test-utils": "^1.3.0",
-    "@vue/vue3-jest": "^28.0.0",
-    "eslint": "^7.32.0",
+    "@vue/eslint-config-prettier": "^7.0.0",
+    "@vue/test-utils": "^2.0.0",
+    "eslint": "^8.16.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-vue": "^9.0.1",
     "jest": "^27.1.0",
     "prettier": "^2.6.2",
     "sass": "^1.52.1",
-    "sass-loader": "^8.0.2"
+    "sass-loader": "^13.0.0"
   }
 }
