@@ -1,12 +1,9 @@
-import Vue from "vue";
+import { createApp } from "vue";
 import Vuex from "vuex";
 import App from "./App.vue";
 import router from "./router";
 
-Vue.config.productionTip = false;
-Vue.use(Vuex);
-
-new Vue({
-  router,
-  render: (h) => h(App),
-}).$mount("#app");
+const app = createApp(App);
+app.use(Vuex);
+app.use(router);
+app.mount("#app");
